fix: stop running gh stats as a background process
This is being removed until a more reliable optimization can be made.
The background process has the potential to stack running processes and
can potenially a gh process running in the background. It also doesn't
really improve the perceived performance because tmux already has an
update interval. In the future a cache should be implemented or
something similar to reduce api calls to GitHub.