refactor: theme system and update styles
- add tmux_get_opt() helper function and replace static theme presets
with a dynamic THEME map driven by tmux options. - update scripts to use
new theme associative array. This will make sure that all the colors
default to the terminals ansi colors so it will alwasy match the users
theme. If the user decides they want to add custome colors, there is a
tmux option to change the any of the 18 ansi colors as well as some
options to change some of the ui elements.